Avoid 100% CPU usage after logback reconfiguration
Commit 73f133f (Add fix for hanging logback thread in TcpAppender) tried to fix
a problem with logback preventing application shutdown because non-daemon
threads were still running by setting the ThreadPoolExecutor's core pool size to
0.

That is problematic since you cannot know whether the app is going to shut down
at all at this point.

We are using logback's "scan" mode[1] which reloads the configuration file if it
changes.

When we have a `TcpAppender` attached to some logger a reload causes a constant
100% CPU usage of a single core.

In light of commit 0d9b01e (Use standalone executor for running TcpAppender
task) this should also not be necessary anymore, because the `TcpAppender` is
using its own `ExecutorService`.

[1]: https://logback.qos.ch/manual/configuration.html#autoScan