Allow TcpAppender to block until dropping events
In its default configuration, logging events are simply dropped when the writer
thread is busy processing an event currently.

There is no way to notice until you realize events are amiss.

This change allows the appender to block for a small amount of time (by default
100 milliseconds) waiting for space to become available.

Also, an informational message will be issued when events are dropped.

The change proposed herewith is basically a backport of the
`AbstractSocketAppender` implementation of logback-classic itself (see [1] and
[2]).

[1]: https://github.com/qos-ch/logback/pull/170
[2]: https://github.com/qos-ch/logback/blob/4dd08e68321548cd85251ec389f5a832505e1268/logback-core/src/main/java/ch/qos/logback/core/net/AbstractSocketAppender.java#L173-L185