Changing log4j depependencies to "provided" scope.
These dependencies are needed in this project to test HEC communication through appenders in different logging frameworks. However, they are not needed at runtime, and in fact are expected to be provided by the consuming app. This change reduces the footprint of the consuming apps by not including log4j as a transitive dependency.

This commit resolves issue #103.