use gson for json serialization
We previously used both gson and json-simple, and the way that we did it meant that
attempting to use a JSON payload would mean having that payload JSON encoded in a
string in the Splunk event. This behavior is retained when using EventBodySerializer
for backwards compatibility, but this should greatly improve the more common case.

use deprecated calls for gson json parser

Spring Boot pulls in a version of gson that does not have the new calls.