HttpEventCollectorLoggingHandler: Default to null for config parameters.
Using empty string as default value parameters, results in JSON events
with empty string as JSON value for the resp. attribute.

This empty string confuses the splunk server and let it reject the http log event.

Use null value as default for all config paramerters as it's done in the
other HttpEventCollector implentations, e.g. HttpEventCollectorLogbackAppender